fix(core): ein eingeschlepptes snprintf ueberschrieb das der C-Bibliothek (MF-523/524)
Der berichtende ASan/UBSan-Volllauf aus MF-517 lief zum ersten Mal ueber
die GANZE Suite statt ueber neun Testnamen. Vier Befunde ausserhalb der
Lecks, drei davon im Produktionscode — und keiner war unter Windows
sichtbar.

MF-523: snprintf war nicht das der C-Bibliothek

src/formats/amiga_ext/snprintf.c definiert GLOBALES snprintf und
vsnprintf — nicht static, nicht umbenannt, ohne Guard ausser _MSC_VER —
und steht in UnifiedFloppyTool.pro. Unter glibc ist snprintf ein echtes
externes Symbol; JEDE Aufrufstelle des Programms band gegen diese Fassung.

Ihr Parser stammt von 1995 und kennt genau:

    -  0-9  l  u  U  o  O  d  D  x  X  s  c  %

Es fehlt `z`. Ein "%zu" wird falsch gelesen, die Argumentliste verschiebt
sich, und ein spaeteres "%s" landet auf einer Zahl:

    SEGV on unknown address 0x28
      #0 fmtstr       src/formats/amiga_ext/snprintf.c:175
      #3 snprintf
      #4 uft_probe_format  src/core/uft_probe_format_impl.c:78

Die Zeile dort formatiert "%zu plugins claim this data at confidence %d;
'%s' wins ...". Gemessen: 10 Aufrufstellen unter src/ benutzen %z, %ll
oder %p — alle zehn standen im ausgelieferten Binary auf diesem Parser.
Im besten Fall falscher Text, im schlechtesten ein Absturz.

Niemand ruft die Amiga-Fassung absichtlich: `plp_snprintf` kommt im ganzen
Baum nur im Kommentar der eigenen Datei vor.

WARUM ES UNTER WINDOWS UNSICHTBAR BLIEB — gemessen, nicht vermutet

`nm` zeigt die snprintf-Symbole des Testbinaries als `t`, also lokal.
MinGWs <stdio.h> liefert snprintf als `static inline` ueber
__mingw_vsnprintf; jede Uebersetzungseinheit bekommt ihr eigenes Symbol,
und die globale Fassung wird nie referenziert. Das Objekt ist trotzdem
gebunden (snprintf.c.obj liegt im Zielverzeichnis).

Die Datei wird jetzt nur noch uebersetzt, wenn UFT_USE_VENDORED_SNPRINTF
ausdruecklich gesetzt ist — fuer eine Plattform, der snprintf wirklich
fehlt. tests/test_snprintf_not_shadowed.c haelt das fest und schreibt
dazu, dass er unter MinGW gar nicht rot werden KANN. Wer das nicht
dazuschreibt, liest spaeter ein gruenes Windows-Ergebnis als Freispruch.

MF-524: die drei uebrigen Befunde

  global-buffer-overflow   src/formats/c64/uft_cmd.c:151 cmd_format
    Die Schleife las name[i] fuer i = 0..15, auch NACH dem Nullbyte.
    `name[i] != '\0'` half nicht — es nahm den else-Zweig und las in der
    naechsten Runde weiter.

  shift exponent 32        src/formats/nintendo/uft_n64.c:391
    #define ROL(i,b) ((i)<<(b) | (i)>>(32-(b))) ist fuer b == 0
    undefiniert, und b kommt aus den DATEN: ROL(d, d & 0x1F).

  left shift of 143 by 24  tests/test_air_cross_validate.c:371
    buf[8] << 24 auf int; ab Byte-Wert 128 undefiniert. Im Test.

WAS DER VOLLLAUF AUSSERDEM GEMESSEN HAT: DEN LECK-RUECKSTAND

LeakSanitizer meldet Lecks in vielen Tests, von 88 Byte bis 8 132 856 Byte
in 321 Allokationen. Der ist NICHT behoben und NICHT beziffert — er ist ab
jetzt sichtbar. Genau deshalb laeuft das scharfe Tor mit detect_leaks=0:
ein ungemessener Rueckstand darf kein Tor bewachen, aber er darf auch
nicht unsichtbar bleiben. Steht als P0-11 in OPEN_ITEMS.

DIE LEHRE

Sieben Speicherfehler hat der eigene Fuzzer gefunden. Diese vier hat er
NICHT gefunden, und keiner davon war unter Windows sichtbar. Ein Werkzeug,
das nur auf einer Plattform geprueft wird, ist auf einer Plattform
geprueft.

VERIFIKATION

  ctest                      243/243 gruen (war 242, +1 neuer Test)
  qmake-Release              baut und linkt, 0 Fehler
  check_consistency.py       25 Kategorien, 0
  ASan/UBSan                 der naechste CI-Lauf ist der Beleg

+++ tests/test_air_cross_validate.c
@@ -363,12 +363,12 @@ static void test_ipf_minimal(void) {
 
     /* Validate CRC of CAPS record */
     uint32_t caps_crc = air_crc32_header(buf, 0, 12);
-    uint32_t stored_crc = (buf[8]<<24)|(buf[9]<<16)|(buf[10]<<8)|buf[11];
+    uint32_t stored_crc = ((uint32_t)buf[8]<<24)|((uint32_t)buf[9]<<16)|((uint32_t)buf[10]<<8)|(uint32_t)buf[11];
     ASSERT_EQ(caps_crc, stored_crc, "IPF CAPS CRC-32 valid");
 
     /* Validate CRC of INFO record */
     uint32_t info_crc = air_crc32_header(buf + 12, 0, 96);
-    uint32_t info_stored = (buf[20]<<24)|(buf[21]<<16)|(buf[22]<<8)|buf[23];
+    uint32_t info_stored = ((uint32_t)buf[20]<<24)|((uint32_t)buf[21]<<16)|((uint32_t)buf[22]<<8)|(uint32_t)buf[23];
     ASSERT_EQ(info_crc, info_stored, "IPF INFO CRC-32 valid");
 
     free(buf);
